Split the block fetch logic into multiple modules
