+++ ouroboros-network/src/Ouroboros/Network/BlockFetch.hs
@@ -1,0 +1,221 @@
+{-# LANGUAGE NamedFieldPuns             #-}
+{-# LANGUAGE RecordWildCards            #-}
+{-# LANGUAGE ScopedTypeVariables        #-}
+{-# LANGUAGE TypeFamilies               #-}
+
+module Ouroboros.Network.BlockFetch (
+    fetchLogic,
+    FetchLogicExternalState(..),
+    --TODO: re-export other bits of public API, mostly client interface
+  ) where
+
+import qualified Data.Map.Strict as Map
+import           Data.Map.Strict (Map)
+import           Data.Void
+
+import           Control.Monad.Class.MonadSTM
+import           Control.Monad.Class.MonadSay
+import           Control.Monad.Class.MonadTimer
+
+import           Ouroboros.Network.Block
+import           Ouroboros.Network.Chain (Point)
+import           Ouroboros.Network.ChainFragment (ChainFragment(..))
+
+import           Ouroboros.Network.BlockFetch.Types
+import           Ouroboros.Network.BlockFetch.DeltaQ
+import           Ouroboros.Network.BlockFetch.Client
+import           Ouroboros.Network.BlockFetch.State
+
+{-
+Let's start with the big picture...
+
+┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄
+Key:  ┏━━━━━━━━━━━━┓  ╔═════════════╗  ┏━━━━━━━━━━━━━━┓   ╔════════════╗
+      ┃ STM-based  ┃  ║active thread║  ┃state instance┃┓  ║ one thread ║╗
+      ┃shared state┃  ║             ║  ┃   per peer   ┃┃  ║  per peer  ║║
+      ┗━━━━━━━━━━━━┛  ╚═════════════╝  ┗━━━━━━━━━━━━━━┛┃  ╚════════════╝║
+                                        ┗━━━━━━━━━━━━━━┛   ╚════════════╝
+┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄
+Notes:
+ • Thread communication is via STM based state.
+ • Outbound: threads update STM state.
+ • Inbound: threads wait on STM state changing (using retry).
+ • These are no queues: there is only the current state, not all change events.
+┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄
+  ╔═════════════╗     ┏━━━━━━━━━━━━━┓
+  ║ Chain sync  ║╗    ┃   Ledger    ┃
+  ║  protocol   ║║◀───┨   state     ┃◀───────────╮
+  ║(client side)║║    ┃             ┃            │
+  ╚══════╤══════╝║    ┗━━━━━━━━━━━━━┛            │
+   ╚═════╪═══════╝                               │
+         ▼                                       │
+  ┏━━━━━━━━━━━━━┓     ┏━━━━━━━━━━━━━┓     ╔══════╧══════╗
+  ┃  Candidate  ┃     ┃   Set of    ┃     ║  Chain and  ║
+  ┃  chains     ┃     ┃  downloaded ┠────▶║   ledger    ║
+  ┃  (headers)  ┃     ┃   blocks    ┃     ║  validation ║
+  ┗━━━━━┯━━━━━━━┛     ┗━━━━━┯━━━━━━━┛     ╚══════╤══════╝
+        │                   │ ▲                  │
+        │ ╭─────────────────╯ │                  │
+░░░░░░░░▼░▼░░░░░░░░           │                  ▼
+░░╔═════════════╗░░           │           ┏━━━━━━━━━━━━━┓     ╔═════════════╗
+░░║    Block    ║░░           │           ┃   Current   ┃     ║ Block fetch ║╗
+░░╢    fetch    ║◀────────────┼───────────┨    chain    ┠────▶║ protocol    ║║
+░░║    logic    ║░░           │           ┃  (blocks)   ┃     ║(server side)║║
+░░╚══════╤══════╝░░           │           ┠─────────────┨     ╚═════════════╝║
+░░░░░░░░░│░░░░░░░░░           │           ┃  Tentative  ┃      ╚═════════════╝
+░░░░░░░░░▼░░░░░░░░░░░░░░░░░░░░│░░░░░░░░   ┃    chain    ┠──╮
+░░┏━━━━━━━━━━━━━┓░░░░░╔═══════╧═════╗░░   ┃  (headers)  ┃  │  ╔═════════════╗
+░░┃    Block    ┃┓░░░░║ block fetch ║╗░   ┗━━━━━━━━━━━━━┛  │  ║ Chain sync  ║╗
+░░┃    fetch    ┠╂───▶║  protocol   ║║░                    ╰─▶║ protocol    ║║
+░░┃   requests  ┃┃░░░░║(client side)║║░                       ║(server side)║║
+░░┗━━━━━━━━━━━━━┛┃░░░░╚═════════════╝║░                       ╚═════════════╝║
+░░░┗━━━━━━━━━━━━━┛░░░░░╚═════════════╝░                        ╚═════════════╝
+░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
+┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄
+
+We will consider the block fetch logic and the policy (but not mechanism)
+for the block fetch protocol client together as one unit of functionality.
+This is the shaded area in the diagram above.
+
+Looking at the diagram we see that these two threads interact with each other
+and other threads via the following shared state
+
+ ═════════════════════════════╤════════════════╤═════════════════
+   State                      │  Interactions  │  {In,Ex}ternal
+ ─────────────────────────────┼────────────────┼─────────────────
+   Candidate chains (headers) │  Read          │  External
+   Current chain (blocks)     │  Read          │  External
+   Set of downloaded blocks   │  Read & Write  │  External
+   Block fetch requests       │  Read & Write  │  Internal
+
+The block fetch requests state is private between the block fetch logic
+and the block fetch protocol client, so it is implemented here.
+
+The other state is managed by the consensus layer and is considered external
+here. So here we define interfaces for interacting with the external state.
+These have to be provided when instantiating the block fetch logic.
+-}
+
+data FetchLogicExternalState peer header block m =
+     FetchLogicExternalState {
+
+       -- | Read the K-suffixes of the candidate chains.
+       --
+       -- They must be already validated and contain the last @K@ headers
+       -- (unless we're near the chain genesis of course).
+       --
+       readCandidateChains    :: STM m (Map peer (ChainFragment header)),
+
+       -- | Read the K-suffix of the current chain.
+       --
+       -- This must contain info on the last @K@ blocks (unless we're near
+       -- the chain genesis of course).
+       --
+       readCurrentChain       :: STM m (ChainFragment block),
+
+       -- | Recent, only within last K
+       readFetchedBlocks      :: STM m (Point block -> Bool),
+
+       addFetchedBlock        :: Point block -> block -> m ()
+     }
+
+-- | The consensus layer functionality that the fetch logic depends upon.
+--
+-- These are provided as input to the fetch logic by the consensus layer.
+--
+data ConsensusFunctions header block =
+     ConsensusFunctions {
+
+       -- | Given the current chain, is the given chain plausible as a
+       -- candidate chain. Classically for Ouroboros this would simply
+       -- check if the candidate is strictly longer, but for Ouroboros
+       -- with operational key certificates there are also cases where
+       -- we would consider a chain of equal length to the current chain.
+       --
+       plausibleCandidateChain :: ChainFragment block
+                               -> ChainFragment header -> Bool,
+
+       -- | Compare two candidate chains and return a preference ordering.
+       -- This is used as part of selecting which chains to prioritise for
+       -- downloading block bodies.
+       --
+       compareCandidateChains  :: ChainFragment header
+                               -> ChainFragment header
+                               -> Ordering,
+
+       -- | Much of the logic for deciding which blocks to download from which
+       -- peer depends on making estimates based on recent performance metrics.
+       -- These estimates of course depend on the amount of data we will be
+       -- downloading.
+       --
+       blockFetchSize          :: header -> SizeInBytes,
+
+       -- | Given a block header, validate the supposed corresponding block
+       -- body.
+       --
+       blockMatchesHeader      :: header -> block -> Bool
+     }
+
+
+-- | Execute the block fetch logic. It monitors the current chain and candidate
+-- chains. It decided which block bodies to fetch and manages the process of
+-- fetching them, including making alternative decisions based on timeouts and
+-- failures.
+--
+-- This runs forever and should be shut down using mechanisms such as async.
+--
+fetchLogic :: forall peer header block m.
+              (MonadSTM m, MonadTime m, Ord peer,
+               HasHeader header, HasHeader block,
+               HeaderHash header ~ HeaderHash block,
+               -- extra debug constraints:
+               MonadSay m, Show peer, Show header, Show block)
+           => FetchLogicExternalState peer header block m
+           -> ConsensusFunctions header block
+           -> FetchClientRegistry peer header m
+           -> m Void
+fetchLogic FetchLogicExternalState{..}
+           ConsensusFunctions{..}
+           (FetchClientRegistry registry) = do
+
+    -- TODO: get this from elsewhere
+    peerGSVs <- newTVarM (Map.empty :: Map peer (PeerGSV (Time m)))
+
+    fetchLogicIterations
+      FetchDecisionPolicy {
+        -- For now, use a fixed policy.
+        -- It's unclear for the moment if this will be fixed external config
+        maxInFlightReqsPerPeer   = 10,
+        maxConcurrentFetchPeers  = 1,
+
+        plausibleCandidateChain,
+        compareCandidateChains,
+        blockFetchSize
+      }
+      FetchTriggerVariables {
+        readStateCurrentChain    = readCurrentChain,
+        readStateCandidateChains = readCandidateChains,
+        readStatePeerStatus      = readPeerStatus
+      }
+      FetchNonTriggerVariables {
+        readStateFetchedBlocks = readFetchedBlocks,
+        readStatePeerStates    = readPeerStates,
+        readStatePeerGSVs      = readTVar peerGSVs,
+        readStatePeerReqVars   = readPeerReqVars
+      }
+  where
+    --TODO: move these next to the FetchClientRegistry
+    readPeerStatus :: STM m (Map peer PeerFetchStatus)
+    readPeerStatus =
+      readTVar registry >>= traverse (readTVar . fetchClientStatusVar)
+
+    readPeerStates :: STM m (Map peer (PeerFetchStatus, PeerFetchInFlight header))
+    readPeerStates =
+      readTVar registry >>=
+      traverse (\s -> (,) <$> readTVar (fetchClientStatusVar s)
+                          <*> readTVar (fetchClientInFlightVar s))
+
+    readPeerReqVars :: STM m (Map peer (TFetchRequestVar m header))
+    readPeerReqVars =
+      readTVar registry >>= return . Map.map fetchClientRequestVar
+
